Fix for conversation roles in UserSearch and ConversationList (#2496)
The bug was caused by using conversation roles when currentConversationId was not yet set - after the first start of the app.
On top of conversation roles we have the team roles. Sometimes the two overlap and that was the source of confusion. For example, in the User Search, we should use team roles to check if the user is able to add services to a conversation, not the conversation roles.

In the process of fixing I made some small changes to `IntegrationDetailsFragment`, but I decided not to rewrite it more than needed. It should be either refactored in another PR, or rewritten in Kotlin completely.