Fix empty conversation list options menu when the current conversation id is not set (#2498)
With conversation roles, we now need to know if the user is the creator of the given conversation to make sure we are allowed to show the "Delete conversation" button. The way it was done involved asking for currentConvId which is not set at the fresh start.
(And which shouldn't be asked for anyway).