Make UserConnectionEvent agree with the backend API
`UserConnectionEvent` has a field for conversation id. On the backend side this field is option -
if it's missing, we can simply convert the `from` user id to RConvId. But the backend provides
the value for this field anyway, so on the Android side this field was required.
The difference was found out only after connecting the Android client to a customized BE which
followed the API to the letter.