Remove old Java code (#2560)
While working on new connect request screens I removed some calls to old Java listeners. One of them became unused in the process. That motivated me to take a closer look to all our Java code. It turned out a lot of it is not used anymore.
Java was used mainly in 2016, before Wire Android moved to Scala. For a long time Scala code used Java as a sort of underlying data transport layer. To this day you can see that Scala fragments implement "containers" which serve to react to events happening in other fragments. Pairs of "listeners" and "observers" are used in other parts for the same purpose. Gradually they were replaced with signals, but they were never completely deleted. In time we dropped a lot of Java classes in favour of Scala and recently of Kotlin, but because we focused on the new functionality, we often just left the unused code be.

This PR removes the old unused code and only in a few places makes changes to the used code, in order to divide a piece of code which has a small part of it used, but the rest can be deleted.