Handle onBackPressed on connect request screens opened with deep links (#2573)
Opening a connect request screen from a deep link may result in a situation when getting to the conversation screen by clicking the back button is impossible. The reason for that is that we jump directly to the connect screen and if the back button relies on the order of initialized screens, one over another, it might fail. Especially the send request screen is at risk because from the app it's always either the Search UI or the Participants screen that is its parent. In the case of a deep link, the parent is the conversation list.

This PR tries to improve the behaviour by popping the stack directly from UntabbedRequestFragment.onBackPressed for deep links. That should take the user back to the conversation list.