A fix for backup restoration (#2577)
fix: Show proper title when asking for password on the backup restoration
fix: make backup password required
fix: Database files after restoration were corrupted.

Currently we show a popup with the same title and message both when we store/encrypt and restore/decrypt the backup. This PR changes the title and clears the message in the latter case, so it won't confuse the user. 
The password is now required. The restoration still takes into account that the backup may not have a password for backward compatibility.
The restoration is now wrapped in a transaction and the account manager's initialization was made lazy, so that it doesn't try to use the database concurrently to the restoration.