Refactor assets to enable image rotation and avoid unnecessary creation of files (#2609)
* Refactor assets to enable image rotation and avoid unnecessary creation of files

This commit is quite big but I felt there was no good way to break it into smaller ones.
With our new assets implementation, we try to treat assets of all types and coming from all sources in the same way as much as possible. In part it was achieved by turning them to input stream - no matter if originally they're files or byte arrays or data streams.
But in certain situations we need to know what we're working on. One such situation is when we send an image: we need to scale it down. Without knowing the path to the original file we were not able to get the EXIF information about the file (not sure why, ask Java) and sometimes the rotation of the scaled-down image came out wrong.
In order to fix it, I would have to carry the EXIF information through many layers of the asset functionality, only to use it if the asset was an image and needed to be scaled down at one point of the whole procedure.
That wouldn't look nice.
On top of that, our original idea of treating everything as input streams meant that sometimes we converted files to input streams and then back to files, copying data, even when using the original file directly would be sufficient, risking performance issues.

So, I decided to refactor assets. Instead of `InputStream` we have now `AssetInput` - a trait which underneath hold the asset as either a real input stream, or a file (or a failure if something went wrong and we just carry information that sending/downloading failed). When we perform an operation on an "asset input" it may either perform a conversion between an input stream and a file, or just access them directly, if the conversion is not really necessary. In case of images from the gallery, which originally contain the EXIF information, the underlying instance is a file, so we can scale it down ad rotate it properly.