fix: Make WSPushService more responsive
`WSPushService` is a class controlling websockets. There's one websocket per account, and they should
be open when the app is in foreground, but closed when it's in background - to save the battery
and prevent Android from killing the app. `WSPushService` also takes care of retrying opening the web
socket if there was an error.

Until now, being offline was also considered an error and `WSPushService` tried (and failed) to reconnect
constantly, with an exponent backoff, i.e. the intervals between tries were getting longer the longer
the device was offline. When eventually the device went online again, the app had to wait until the end
of the given interval to try again.
Also, because of implementation details of how signals work, `WebSocketService` sometimes didn't notice
that a reactivation is necessary. Together these two errors resulted in delays of receiving events
from the backend, sometimes so long that for the customer it looked as if we missed some events.

The PR fixes the way WebSocketService handles activation and deactivation of `WSPushServices`, and
it introduces the change in the network mode (going online/offline, but also a switch between different
networks) as one more thing that can trigger activation and deactivation. `WSPushService` should be
more responsive right now, and the app may also become a bit more battery-friendly, as it does not
try anymore to reconnect when offline.

I also refactored a bit of code in other places where the information about being online/offline is
needed. In a few of them we tried to get that information by asking the signal for `.currentValue`,
which is less reliable than `.head`.