fix: A race condition when a push notification results in a new message
A new message in the messages storage triggers sending back a read receipt and possibly some other updates, e.g.
in the badge of unread messages. If one of those processes results in a sync job, it is possible that it will get
in the way of processing the original push notification.
This PR fixes this by stopping the event streams of of `onAdded`, `onUpdated`, and `onDeleted` in `MessagesStorage`.
The events which should be published to the streams are queued and released only after processing of the original
push notification. It should help us with some cases of freezing at start and delayed/missing notifications.

Please note that one other big problem with push notifications is that Android does not wake our FCMHandlerService.
That one can be fixed probably only by updating FCM, which right now is impossible. As part of this PR I updated
our dependencies a bit, but it's not enough.

FCM is bumped to 20.1.0.
There's some slight refactoring too, aimed at investigation why sometimes FCM does not work when the app is off.