chore: OkHttp update to 3.14.6
The jump to 3.14 means that we need to add source and target compatibility with Java 8 in app.gradle.
We break the backward compatibility with Android 4.
In return, we get a lot of bugfixes and better cooperation with the websocket, i.e. more reliable
messages sending.

The whole list of changes:
https://square.github.io/okhttp/changelog_3x/