refactor: Disable tracking on the TrackingService level
We don't track events right now, but we have a scaffold functionality left which can be used again if we decide to use another
tracking system in the future. Right now, calling a track method on an event results only in a log message.
However, before that happens, in some cases `TrackingService` and `GlobalTrackingController` can perform quite a lot of
operations, esp. `TrackingService.contribution` and `TrackingService.assetContribution`.
I decided to move the code around a bit to prevent that. With this PR if the preference  `GlobalPreferences.AnalyticsEnabled`
is set to `false` (which it is in production) we will skip performing these operations.