perf: Optimize sharing many assets to many conversations
If the user goes to, for example, the gallery, and chooses a few photos and then share it to Wire to a few conversations at once,
we handle that request in `SharingController`. We go through the collection of uris we got and for each uri we go through
the collection of conversations, and for each pair we call a method which sends that one asset to that one conversation.
This PR inverts the procedure - first we go through conversations, and then through uris - and provides functionality for sending
many assets at once to one conversation. Thanks to that, some operations, like setting the conversation's "last read time" or
setting it to be the current conversation, are performed only once per sending all assets to that conversation, not each time for
each asset. That makes the whole process much quicker, and also helps us avoid possible race conditions due to quick switching
of the current conversation (still to be investigated).

Connected to: https://wearezeta.atlassian.net/browse/AN-5917