fix: Handle all exceptions when creating tones for Wire's notifications (#2633)
* fix: Handle all exceptions when creating tones for Wire's notifications

When we open Wire, we try to register our own custom sounds for notifications: one for new message, and one for a ping.
(Right now we're doing it every time we open the app - maybe we don't have to do it so often)
In order to do that we have to find the external notifications folder, find a tone file within it, copy our sound to that file, and notice Android about it by inserting some data into the content resolver.
At each step of this process something may go wrong and an exception will be thrown. We try to handle those exceptions, but it seems that at least on some phones with Android 10 one of the exceptions escapes. (You can look into Google Play Console for that).

This PR refactors the process, wrapping everything in functional "Tries" in Scala. This way we should be able to catch all exceptions and handle them gracefully.