fix: Count the number of call participants with ConversationsService.activeMembersData (#2646)
* fix: Count the number of call participants with ConversationsService.activeMembersData

The video call is available only in group chats with 4 or fewer participants. If the current user belongs to the team and the number of participants is ok, a video icon will appear in the header. However, the way it was implemented
meant that it was possible that when a new participant was added and then removed (so the group size would grow to 5 and then get back to 4), the video icon would disappear and not appear again. Also, when the current user was called from another device, they could not join the video call.

The PR changes the way we count call participants and refactors the calling code a bit.

* a fix to an error in counting unverified users