fix: The sync after backup restoration shouldn't be treated as the first sync.
This is another fix to the backup restoration issue. The restoration itself should now be done successfully,
but we still have problems with setting up the app properly afterwards. One of them is that the first sync
with the backend after the backup restoration is performed as if this was a fresh install. In such case,
the app doesn't ask for old notifications which should be done in case of the backup restoration.

Unfortunately, this is not the only issue we have. The current workaround is to restart the app just after
the backup restoration.