fix: A fix to IndexOutOfBoundsException in PagedStorageDiffHelper
This happens because of the AndroidX bug, described here:
https://stackoverflow.com/questions/55546116/room-livedata-with-paging-getting-indexoutofboundsexception-in-some-devices
The bug is currently the main source of crashes, according to Google Play Console.

The suggested fix is to update the androidx.paging library to 2.1.1 (from 2.0.0) and to disable null placeholders
when building a paged list config (in MessagePagedListController).

Testing: As far as I understand, disabling null placeholders may mean slower performance of scrolling through
the messages' list in the conversation or may result in some glitches during the scrolling (can someone confirm this?).
Please go to a long conversation and scroll fast up and down.
Take note than on a dev build the scrolling may display some performance problems anyway, so in case of glitches
please compare with a dev build without this fix.