refactor: Automate turning collections of Identifiables into maps
I made a small Scala exercise on the side and was able to hide one common operation under the hood.

In Scala code we have many classes which implement `Identifiable[K]`. It means that in the given class
there is one or two (and in theory more) fields which serve as a unique identifier, and then K is
the type of that field, or a tuple of them both. Usually those classes are the ones with corresponding
db tables (UserData, ConversationData, ...) but it's not limited to them.

In many places in our code we retrieve a collection of instances of one of those classes and then turn
it into a map where keys are identifiers and values are the instances themselves. It's a bit redundant,
but it helps in performance and readability. With help from the Polish Scala slack channel I was able
to put that transformation in one implicit (i.e. extension) method, `toIdMap`.

It's just a small refactoring thanks to Scala type system magic. It's low priority. Feel free to just
look at the code and ask questions if you're interested, and we can either merge it or close it later on.