feat: Delete buttons when deleting messages (#2681)
In many cases we delete messages by its id, without checking its type. I decided that in such situations it can be actually faster to just call a method on the Buttons table which will do nothing than querying the Messages table for the message type and then calling the method on Buttons only if the type is Composite.

I moved `def addButtons` from `MessagesService` to `MessagesContentUpdater`. This makes it clearer that buttons belong to a message and fewer changes in the tests are needed to support that.

I also took this opportunity to add a few checks for "empty" in storages. SOmetimes it happens that we call a method on the database with an empty set of identifiers, so the method will do nothing, but anyway it will first call a bunch of other methods and run a query. It's better to avoid it.