fix: The contents go to a wrong conversation (#2696)
fixes https://wearezeta.atlassian.net/browse/AN-5917

If the sync with BE takes a long time it is possible that in the meantime the user will open a conversation,
read something, and the go to another one. In such case the contents of the conversation will not be refreshed (so, in fact, the conents does not "go" to a wrong conversation - it just does not go away at all).
The reason was one line of code which I introduces in September 2019 during one of attempts to find out why there are delays in syncing. The line stops the contents from refreshing until the sync is finished.
Since then we figured out some other ways to prevent delays, and this one introduced the bug, so I just removed it.

Anyway, the delays still happen. More work is needed.