fix: Ignore incoming Button Action event
If there is a poll in a group chat and the other user responds, the current user receives a Button Action event
which should be ignored. By default, we don't ignore unknown message events, but treat them as "unknown" and
store them. The side-effect of that is that an empty message appears in the conversation which indicates that
the client don't know how to display a message, but anyway the message is there. This may happen for example
when a new version of Wire introduces a new message type (like now, with Composite), but that new type of
message is received on a client with an older version.