fix: Postpone processing notifications until after the sync
The processing and displaying of notifications seems to be a major part of very slow sync with BE.
This PR fixes a bug in `NotificationService` which causes the process to build and display the same
notification many times by postponing this process until processing of incoming events is done.

Imagine the user goes offline, three messages are sent to them, and then they go online again.
The sync starts and all three messages are received. Three notifications should be created and
displayed. But without this PR this process goes like that:

Message #1 (M1) comes -> Notification #1 is created (N1) -> Notification #1 is displayed
M2 comes -> N1 is dismissed -> N1 and N2 are created -> N1 and N2 are displayed
M3 comes -> N1 and N2 are dismissed -> N1, N2 and N3 are created -> N1, N2, and N3 are displayed

The process is done n times for n new messages, and within it the notifications are dismissed
and created (n + (n-1) + (n-2) + ... + 1) times. The creation of a notification is also slow and
can be improved a bit - but in another PR.

With this PR, the process runs only once, after all messages come. All notifications are created once,
and none of them need to be dismissed.