fix: Check if the poll's author is in the conversation
fixes https://wearezeta.atlassian.net/browse/AN-6763

If the author's of the poll is no longer a member of the conversation, responding to the poll should result in the error message.
But a button action (i.e. a button click) is sent to the Backend as a regular message, so the standard error handling procedure
doesn't work for us here - we receive a successful response from the Backend that the message was received.
Instead, before sending the button click we have to check who is the author, check if that person/service is still an active
member of the conversation, and send it only if it's true. Otherwise, the error message should be displayed.