fix: Resend messages edited when offline
fixes https://wearezeta.atlassian.net/browse/AN-6703

After the user edits a message, the app tries to send it immediately, and if it fails, it changes
the state of the message to FAILED. That should mean that in the conversation the message footer
should display the "Resend" link so that the user can try again. But there was a bug: the fact
that the message was edited had priority so the footer displayed the "Edited: [time]" info, and
without the "Resend" link it was not possible to resend the message. The PR fixes that.

A different problem arises when the user tries to edit a message which was never send. Then,
when the message is resent, it carries to the receiver the information that it's an edited version
of an original message, but the original message is not found on the receiver's side, so the
edited version is ignored (this is by design). To make it consistent with other platforms,
we decided to simply prevent the user from editing unsent messages. I removed the "edit" button
from the list of possible actions in such case.

Note that the PR does not implement automatic resending of message when the device goes online.
We could do it, but that's a new functionality really, not just bug fixing. Another ticket will
address it.