fix: Display websites' previews on Android 9-10 (#2731)
fixes https://wearezeta.atlassian.net/browse/AN-6685

Since version 9, Android by default disallows cleartext traffic and if the website uses that  we get an exception:
```
java.net.UnknownServiceException: CLEARTEXT communication to heise.de not permitted by network security policy
```
One way to fix it would be to disable this security check and allow cleartext again, but I don't want to do that. Instead, I decided to handle that specific exception and if it happens, retry the request with "https://" instead of "http://". In many cases that will  be enough. In the others, it simply means that the problem is on the website's side and we won't show the preview.
The weblink itself is not affected - if the user clicks on it, the webpage starting with "http://" will be opened.