feat: Remove MessageIndexStorage to assist migration (#2722)
feat: Remove MessageIndexStorage to assist migration

MessageIndexStorage and its virtual db table MessageContentIndex served as text cache for searches. There was a conflict with the current setup of the virtual table and our migration of the database to Room. Instead of trying to somehow fix the problem while maintaining the compatibility for migration from the previous versions of Wire, I removed MessageIndexStorage and moved the searching MessagesStorage and MessageData.