fix: The Buttons table created by Kotlin + the new migration
The Buttons table was created by Scala and then used by Room which could leadto a conflict:
Room expected the primary keys numbers to be different that they were. The solution for that
is from now on disable Scala from creating nwq tables. They should be created on the Kotlin
side and it's Scala which should conform to the schema.

Another change is the introduction of a new migration (127 -> 128) which repeats changes
we made to the schema since the last release. If at least in some cases the crash is not due
invalid schema but due to an error in copying data (e.g. copying a null to a non-null column)
then this may fix the problem.