Room nullables: What can actually be null from Scala? (#2767)
* Updated tests to reflect true values in the entities

* fix: The Buttons table created by Kotlin + the new migration

The Buttons table was created by Scala and then used by Room which could leadto a conflict:
Room expected the primary keys numbers to be different that they were. The solution for that
is from now on disable Scala from creating nwq tables. They should be created on the Kotlin
side and it's Scala which should conform to the schema.

Another change is the introduction of a new migration (127 -> 128) which repeats changes
we made to the schema since the last release. If at least in some cases the crash is not due
invalid schema but due to an error in copying data (e.g. copying a null to a non-null column)
then this may fix the problem.

* fix: Default values, a new migration, fixed "verified" fields (#2772)

* fix: Default values, a new migration, fixed "verified" fields

1. Each *Entity was given default values for non-null fields.
2. Migration code was moved to a new migration. The migration (126, 127) is empty now.
All the code is in (127, 128). This way people migrating from 126 will just skip 127
and all the work will be done in 128, and for people who already are on the faulty 127
we may have a fix this way.
3. "INSERT INTO ..." clauses from the migration were changed to "INSERT OR IGNORE INTO ...".
I tried to use "ON CONFLICT REPLACE" instead but it seems Android SQLite does not
support this syntax. "INSERT OR IGNORE" means that if a conflict occurs then we skip the entry.
We may lose some data this way, but I suspect this is very rare and it's a much more
important not to crash on migration.
4. There are two fields called "verified" in ConversationData and UserData. For some reason
making them non-null with a default value breaks the schema. I suspected the problem lies
on the Scala side, but I fixed that and the problem still occured, so I made the fields nullable.

It looks like a lot but it's really not. Most of the changed lines of code come from
moving code from one migration to the other.

* supress detekt errors

* Fixing Migration unit tests

* Fixing conflicts and an error in PushNotificationEvents, removing duplicated tests

Co-authored-by: Maciej Gorywoda <makingthematrix@gmail.com>