fix: Global database changes moved to a new migration
This is similar to what we did with the user database: all changes since the 3.47 version with the bug in migrations
were moved to a new migration, so that people with the faulty 3.47 may have a chance for a fix.
Please test by both going back to 3.46 and updating to this one, and by going back to the faulty 3.47 and updating.

There's also one small change to the Buttons table migration as I ran into a problem with it while working on this fix.
It's possible that the Buttons table schema created by the faulty 3.47 is different than what we have now.
It can be fixed by simply deleting the old table - Buttons is a new feature and the table should be empty.