fix: Exact match search (#2812)
There was a bug in the exact match searching - we tried to parse the response from BE as wrong case class. The bug was there even before Large Teams. The right way to get the exact match is very similar to how we look for users by ids, so I decided to move that client functionality from UserSearchClient to UsersClient.