fix: Show an external/partner user in the Add to Group screen
fixes https://wearezeta.atlassian.net/browse/AN-6876 but also fixes the opposite case.
The external user should be visible in the search to the "invitor" before the invitor types in a query,
but the external user should NOT be visible to non-invitors in the same situation.
Only when a non-invitor types in an exact match search, the external user appears in the result.
The behaviour is the same for normal user search and for Add to Group user search.