refactor: Remove NameUpdater and change the way we choose names for conversations
This is ground work for better handling of removed users after the Large Teams feature.
We have three main types of conversations: 1-to-1, fake 1-to-1, and group. Group conversations
should have names of their own, but some older ones don't. For those which don't, and for
all other types, we should use names generated from the list of members.
Until now, we used a field `ConversationData.generatedName` to store the generate name.
The name had to be updated every time one of the conversation members changed their name.
A special service, `NameUpdater`, did that in response to every name changing event.
The name was also updated for real, new group conversations which have their own names, so generated
names were never used. And on top of that we had messy code in several places in the app
which sometimes decided the conversation name should be changed to "Default" when it really
shouldn't.

This PR removes `NameUpdater` and makes `generatedName` obsolete. Instead, we will have one place,
in `ConversationsService` where we create conversation names on the fly. We will use "Default"
only when there is no way to generate the name.

Tests:
1. Create a number of 1-to-1, fake 1-to-1, and group conversations.
2. Change names of members in the conversations:
2a. The names of the real group conversations should stay the same.
2b. The names of 1-to-1 and fake 1-to-1 conversations should change.
3. Remove users from teams.
3a. Conversation names should stay the same although user names can change to "Default"
(this will be fixed in another PR)
4. If you have a big conversations list, test if scrolling didn't slow down (in comparison to `develop`)