fix: Too many prekeys requests
With the Large Teams feature we changed the way we specify to what users should a message be sent. That led to error responses
from the bakend in some cases which in consequence trigger new prekeys requests. There are way too many prekeys requests
coming from Android devices to the backend right now - basically at least one for every message received by a device with
the new version if the user also has Wire on another device.

The message types which are affected are: last reads, delivery confirmations, and read receipts. In standard cases we want to
send the message to all users in the conversation - and that works well - but in these three cases (*) we want them to be sent
only to other self devices and to devices of the sender of the original message. So, we put those user ids on the list of
explicit recipients and later we put that list in the `report_missing` field of the OTR message. And here's something unclear
starts to happen. In those three cases, sending the message to the conversation (`/conversations/[conv id]/otr/messages` endpoint)
results in the error 412, Clients Mismatch with the data that a client is missing. In consequence, the Android app requests
new prekeys for that client.
In cases of 1:1 conversations, that client was another device of the current user. If the current user was on the list of
recipients, the error didn't happen, but in those three cases it wasn't there. Also, in case of group chats, the error happened
and I'm still not sure why.
Before Large Teams, the error didn't happen, because the list of recipients wasn't put in the `report_missing` field
at all (which means that the message is send to everyone in the conversation). So, it looks like the error happens when
`report_missing` specifies the list of recipients, but the backend decides for some reason that the list is wrong.

In this fix, I change the list of recipients for last reads, delivery confirmations, and read receipts, according to what
I know how it should work, but I also revert to an empty `report_missing` field because even after those changes the error
was still happening. We need a discussion about it and more tests. In the meantime, this may work as a hotfix.

I also made some indentation and other small changes to improve readability. It helps me think.

*) There are also button confirmations - but they go only ot the poll's owner, not to the other self devices, and there are
ephemeral expirations, but they seemed to work well.