backend-api-proto bump to 2.3 + use NewOtrMessage.report_missing
It turned out the real reason was that Android uses an obsolete version of backend-api-proto.
The new version was compiled and put on Bintray, and the version number in Android was bumped.
In the new version we have the reportMissing field in OtrNewMessage. Our report_missing from
OtrMessage has to be copied there.