fix: Too many prekeys requests (#2833)
* fix: Too many prekeys requests

With the Large Teams feature we changed the way we specify to what users should a message be sent. That led to error responses from the bakend in some cases which in consequence trigger new prekeys requests. There are way too many prekeys requests coming from Android devices to the backend right now - basically at least one for every message received by a device with the new version if the user also has Wire on another device.

The message types which are affected are: last reads, delivery confirmations, and read receipts.   In those three cases, sending the message to the conversation (`/conversations/[conv id]/otr/messages` endpoint) results in the error 412, Clients Mismatch with the data that a client is missing. In consequence, the Android app requests new prekeys for that client.
Before Large Teams, the error didn't happen, because the list of recipients wasn't put in the `report_missing` field at all (which means that the message is send to everyone in the conversation). 

It turned out the real reason was that Android uses an obsolete version of backend-api-proto.
The new version was compiled and put on Bintray, and the version number in Android was bumped. In the new version we have the reportMissing field in OtrNewMessage. Our report_missing from OtrMessage has to be copied there.