fix: Re-add hanfling of team.member-join events
fixes https://wearezeta.atlassian.net/browse/AN-6896

At one point in the Large Teams feature implementation we removed the code for handling incoming events
of new members joining the team. Later design changes reverted that decision but the code was not
reintroduced, resulting in that, for example, new team members are initially invisible in SearchUI.

This PR reintroduces a few lines which sync new team members if the event comes.