fix: Update conversation names in real time (#2846)
fixes https://wearezeta.atlassian.net/browse/AN-6903
fixes https://wearezeta.atlassian.net/browse/AN-6895

After the recent change in how we decide on conversation names, the conversation list adapter stopped responding
to those changes in real time, at least in some cases:
1. After starting the app, the adapter displays "Default" as the conversation name before the storage is initialized. This should be changed to the actual name just after initialization, but the adapter was unresponsive (AN-6903).
2. If we get a user-update event with the new name, the data changes, but the adapter does not update (AN-6895).

This PR fixes both by adding a listener to the conversation name signal in the conversation list row. It's not the ideal solution but it's consistent with how conversation list rows update other data. I think that the underlying problem is that the conversation list row is right now a very complex entity which breaks the design pattern (it's a row in an adapter but it updates itself through signals instead of using the adapter). I don't want to refactor it now so I leave it as it is, but I think we should rewrite it in the future.