fix: Update user data on opening the user profile (#2857)
fixes https://wearezeta.atlassian.net/browse/AN-6866

It turned out that the most common way of opening a user profile - from the conversation - didn't sync the user. There is a lot of ways to open participant's profiles, depending on that if we are connected to that user, and from what screen we are opening the profile. The code is very complex. I managed to narrow down the paths where we should sync the user before opening the profile to two: one in ConversationManagerFragment (this is when we
open the profile from the conversation) and one in ParticipantFragment (when we go from other places). I decided that trying to merge into one them would make the code too unreadable.