fix: System messages on the user being removed/leaving the conversation or team
fixes https://wearezeta.atlassian.net/browse/AN-6859
fixes https://wearezeta.atlassian.net/browse/AN-6909

In the past, when a user was removed from a conversation, a MemberLeaveEvent should have come
and only in the response to that event we should have displayed a system message "User left"
or "User was removed" or "You removed user" or "Another user removed user".
Now the situation became more complex: We can no longer rely on events and a user may be removed
from the conversation because they were removed from the team entirely (which we can discover
either because of another event type, MemberEvent.MemberLeave, or in a few different ways.

This PR ensures that we display system messages when we discover that the user left or was
removed from the conversation in any of those ways and also tries to find out if we don't
do it more than once.