fix: When looking for a username of a removed user, we should get the old name instead of "Default"
fixes https://wearezeta.atlassian.net/browse/AN-6873

I decided to put this PR on top of the last_system_messages branch instead of develop because they both
contribute to fixing the bug.
Please note that this works because instead of removing the user from the database, we often just mark
them as deleted. If the user is really missing from the database for some reason, and the app is restarted,
so that the name is also not in `UserService.userNames` which serves as a cache, the system message
will display "Default".