Deleted conversation names (#2863)
fix: Preserve conversation names when the last user leaves

Right now we decide on conversation names basing on either that they have their own names or on usernames of its members. This creates a problem with fake 1:1 conversations, and some old group conversations, when the other user leaves. We no longer have information about the other member of the conversation, but also the conversation doesn't have its own name. On the other hand, we don't want to just set the name of the conversation for every such conversation, because it's more intuitive if the conversation name constructed of user names changes along with the user names.