fix: Open the profile from a mention (#2872)
fixes https://wearezeta.atlassian.net/browse/AN-6908

In the Large Team feature we refactored how a single participant profile is being opened. We left out that it's possible to open it also from tapping a mention. It almost worked by itself, but the selected participant was not set which led to displaying a blank screen - the adapter didn't know where to look for contents.

Also, I removed the syncParticipant method from SingleParticipantFragment - syncing is done differently now.